requirements.txt
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
beautifulsoup4==4.12.3
boto3==1.35.97
botocore==1.35.97
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.6
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.8.2
jmespath==1.0.1
lxml==5.3.0
numpy==2.2.1
openai==1.59.3
packaging==24.2
pandas==2.2.3
patsy==1.0.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
statsmodels==0.14.4
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
